+++ fancypages/api/serialisers.py
@@ -1,12 +1,11 @@
 from django.db.models import get_model
 from django.utils.translation import ugettext as _
-from django.forms.models import modelform_factory
 from django.template import loader, RequestContext
 
 from rest_framework import serializers
 
 from .. import library
-from ..dashboard import forms
+from ..library import get_block_form
 
 FancyPage = get_model('fancypages', 'FancyPage')
 ContentBlock = get_model('fancypages', 'ContentBlock')
@@ -41,12 +40,7 @@ class RenderFormFieldMixin(object):
         }
 
     def get_form_class(self, obj):
-        # check if the block has a class-level attribute that
-        # defines a specific form class to be used
-        get_form_class = getattr(obj.__class__, 'get_form_class', None)
-        if not get_form_class:
-            return modelform_factory(obj.__class__)
-        return modelform_factory(obj.__class__, form=get_form_class())
+        return get_block_form(obj.__class__)
 
 
 class BlockSerializer(RenderFormFieldMixin, serializers.ModelSerializer):
@@ -68,17 +62,7 @@ class BlockSerializer(RenderFormFieldMixin, serializers.ModelSerializer):
         return super(BlockSerializer, self).restore_object(attrs, instance)
 
     def get_form_class(self, obj):
-        model = self.object.__class__
-        get_form_class = getattr(model, 'get_form_class')
-        if get_form_class and get_form_class():
-            return modelform_factory(model, form=get_form_class())
-
-        form_class = getattr(
-            forms,
-            "%sForm" % model.__name__,
-            forms.BlockForm
-        )
-        return modelform_factory(model, form=form_class)
+        return get_block_form(self.object.__class__)
 
     class Meta:
         model = ContentBlock